refactor: domain 추가에 따른 폴더 구조 변경
