+++ src/main/java/org/programmers/springorder/voucher/service/VoucherService.java
@@ -1,9 +1,9 @@
-package org.programmers.springorder.service;
+package org.programmers.springorder.voucher.service;
 
-import org.programmers.springorder.dto.VoucherRequestDto;
-import org.programmers.springorder.dto.VoucherResponseDto;
-import org.programmers.springorder.model.Voucher;
-import org.programmers.springorder.repository.VoucherRepository;
+import org.programmers.springorder.voucher.dto.VoucherRequestDto;
+import org.programmers.springorder.voucher.dto.VoucherResponseDto;
+import org.programmers.springorder.voucher.model.Voucher;
+import org.programmers.springorder.voucher.repository.VoucherRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
