+++ src/test/java/org/programmers/springorder/model/VoucherTypeTest.java
@@ -2,6 +2,7 @@ package org.programmers.springorder.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.programmers.springorder.voucher.model.VoucherType;
 
 import java.util.InputMismatchException;
